move sqs to its own folder
