+++ sqs/plugin.go
@@ -5,6 +5,7 @@ import (
 	"github.com/spiral/roadrunner-plugins/v2/internal/common/jobs"
 	"github.com/spiral/roadrunner-plugins/v2/jobs/pipeline"
 	"github.com/spiral/roadrunner-plugins/v2/logger"
+	"github.com/spiral/roadrunner-plugins/v2/sqs/sqsjobs"
 	"github.com/spiral/roadrunner/v2/events"
 	priorityqueue "github.com/spiral/roadrunner/v2/priority_queue"
 )
@@ -31,9 +32,9 @@ func (p *Plugin) Name() string {
 }
 
 func (p *Plugin) ConsumerFromConfig(configKey string, e events.Handler, pq priorityqueue.Queue) (jobs.Consumer, error) {
-	return NewSQSConsumer(configKey, p.log, p.cfg, e, pq)
+	return sqsjobs.NewSQSConsumer(configKey, p.log, p.cfg, e, pq)
 }
 
 func (p *Plugin) ConsumerFromPipeline(pipe *pipeline.Pipeline, e events.Handler, pq priorityqueue.Queue) (jobs.Consumer, error) {
-	return FromPipeline(pipe, p.log, p.cfg, e, pq)
+	return sqsjobs.FromPipeline(pipe, p.log, p.cfg, e, pq)
 }
