+++ sqs/sqsjobs/config.go
@@ -1,4 +1,4 @@
-package sqs
+package sqsjobs
 
 import "github.com/aws/aws-sdk-go-v2/aws"
